template update

+++ forms.py
@@ -6,4 +6,4 @@ from wtforms.validators import InputRequired
 class ConverterForm(FlaskForm):
     base_currency = StringField("Converting From", validators=[InputRequired()])
     converted_currency = StringField("Converting To", validators=[InputRequired()])
-    amount = FloatField("Amount", validators=[InputRequired()])
+    amount = IntegerField("Amount", validators=[InputRequired()])
